pyproject.toml
[project]
name = "restaurant-name-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp==3.9.3",
    "ipykernel>=6.29.5",
    "langchain==0.1.6",
    "langchain-community==0.0.19",
    "langchain-core==0.1.22",
    "langchain-google-genai>=1.0.1",
    "python-dotenv==1.0.1",
    "streamlit==1.31.0",
]
